Move dev dependencies to dependencies
Installing NPM packages from GitHub means we don't have the benefit
of NPM running our babel/prepare scripts on publish.

When you try to install Tailwind, it'll blow up because it tries to
run the prepare script on install and none of the dev dependencies
are installed.

There may be a better way to handle this but there is zero fucking
information on the internet about what a "proper" private package
development workflow looks like, especially when you have a build
step.